Add confirm modals in save and discard operations

+++ src/main.js
@@ -106,11 +106,19 @@ function animate(time) {
 }
 
 function save() {
-    localStorage.setItem('bestBrain', JSON.stringify(bestCar.brain));
+    const isToSave = confirm("This will update the car's brain. Are you sure you want to do that?");
+
+    if (isToSave) {
+        localStorage.setItem('bestBrain', JSON.stringify(bestCar.brain));
+    }
 }
 
 function discard() {
-    localStorage.removeItem('bestBrain');
+    const isToDelete = confirm("This will update the car's brain. Are you sure you want to do that?");
+
+    if (isToDelete) {
+        localStorage.removeItem('bestBrain');
+    }
 }
 
 /**
